Assignment-1/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
boto3==1.37.0
botocore==1.37.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jmespath==1.0.1
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-multipart==0.0.20
requests==2.32.3
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
